requirements.txt
# ============================================================
# NTL-SysToolbox — Dépendances Python
# Compatibilité : Python 3.9+
# ============================================================

# Lecture de la configuration YAML
pyyaml>=6.0.1

# Connexion WinRM vers serveurs Windows
pywinrm>=0.4.3

# Connexion SSH vers serveurs Linux
paramiko>=3.4.0

# Connexion et dump MySQL
mysql-connector-python>=8.3.0
